Update components to align better with our infrastructure
As we have migrated our infrastructure, there have been some changes to our components. The biggest/most important change is that we now use Apicurio for our schema registry, instead of the Confluent Schema Registry. In practice this only means that we need to add /apis/ccompat/v7 to the address.

Another "big" change here is that I've altered the sample CLI producer and consumer to use python instead of the confluent console schema proucers/consumers. The reason is that it should be a bit closer to quick demo/hackaton code while still being easy and short to inline into a container/service.

Moreover, the brokers now use the plain apache/kafka image from dockerhub.