Chunking workshop: Add first basic example
ToDos at this point are cleaning it up a bit, maybe making it a tad more
realistic by adding an api call that allows you to submit payloads
yourself (the baked in cat ipsum is ok, but not great), adding support
for fetching blobs ad hoc by id (keep metadata in dict, fetch chunks
when asked), verifying the payload schema (for the example it could be
as simple as there being one string or byte array field?), making the
code structured into a bit smaller pieces that can be implemented one
after another in a ws setting. Probably more. But, this works and will
do for now.